app/requirements.txt
asgiref==3.3.4
certifi==2021.5.30
click==8.0.1
colorama==0.4.4
fastapi==0.65.2
h11==0.12.0
importlib-metadata==4.5.0
joblib 
pydantic==1.8.2
scikit-surprise==1.1.1
scipy==1.6.3
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.14.0
wincertstore==0.2
zipp==3.4.1
pydantic
